Made some major changes to make the terminology more consistent. I was using the terms provider, implementor, and module in directory/file names, documentation, method and function names, and variable names...all to mean the same thing. So I decided to standardize on module (and variations thereof).
Also, did extensive work on the ZkLoader class, the lib/auth/load.php
script, and so forth. Just started work on lib/logging/services.php, too.
The core of the work in this commit is the code required to automagically
load the zookeeper core and modules without having to hard code all their
names (etc) into ZkLoader, etc.

Seriously need to document the way ZkLoader works and the convention for
naming, creating, etc services (lib/*) and their corresponding modules.

I have decided that the Zookeeper core and modules are going to use the
Zookeeper error handling service to report errors. This is because the
functionality provided by the error handling service is so core and central
to the needs of the rest of the Zookeeper code. This is similar to the fact
that some of the Zookeeper core and modules will require session handling
to be enabled (such as the auth service).

Lastly, created the file ZkFunctions.php that will contain miscellaneous
functions used by the zookeeper core. This file is NOT NOT NOT a place
to put code specific in much detail to any of the services or modules,
as that code (if needing to be shared) should probably be incorporated
into the service itself. Again, this file is only for reusable functions
specific to the Zookeeper core. It will probably remain relatively small.