Changed the default behaviour to use Marc's experimental cpde... note this is devel, and would be great to get some field testing (courier/uw would be great)... it seems to fix the folder display issues with Exchange IMAP where folders where appearing in the wrong order.
Fixed an E_ALL error, and pulled the $imapConnection into the function,
Marc are you running with E_ALL?  The code caused an immediate error as it
couldn't connect to the IMAP server, causing a nice folder order to appear:
  (http://nuts.netdork.net/bad_folders.jpg)

Also fixed a couple of cases of "Editor Envy" and replaced \t with "    "
(hence the massive diff).

Also fixed the code so that the "Trash" is now selectable, and moved the
 (purge)
link back to after the mailbox name.