Fix for moving last message in a folder. Rather than change how messages are indexed, this confines changes to read_body ...
move_messages.php is used to move the last message in the index.
This is the same function that is used when moving messages from the
list index.

If there is a "next" message, read_body deletes the message itself..
(theoretically, we could use move_messages here, too, and just change
the resulting location.... will have to think about that.. )