Another big update:
1) do not do a search 1:* UID request in case we do not sort. That isn't
needed, just use a non uid fetch and request the UID together with the rest
of the headers.

2) Started on adding envelope parsing to sqimap_get_small_header list so we
can remove the imap related code from the messages class.

2) cleaned up the functions in mailbox_display, don't use trillions of
arguments when you can provide an array as well.

3) adapted right_main.php and search.php to fit the adapted functions in
mailbox_display

I hope the code makes more sense now an is easier to understand so that we
can start on separating the logic from the html output and prepare for
templates. Volunteers?